# Speed up `deprecate_soft()` / `deprecate_warn()`, actually warn after 8 hours, and interpolate `({arg})`