--- a/R/signal.R
+++ b/R/signal.R
@@ -24,7 +24,7 @@
 #' }
 #' foofy(1, 2, 3)
 signal_stage <- function(stage, what, with = NULL, env = caller_env()) {
-  stage <- arg_match(stage, c("experimental", "superseded", "deprecated"))
+  stage <- arg_match0(stage, c("experimental", "superseded", "deprecated"))
   what <- spec(what, env = env)
 
   if (is_null(what$arg)) {