--- a/R/signal.R
+++ b/R/signal.R
@@ -28,9 +28,9 @@ signal_stage <- function(stage, what, with = NULL, env = caller_env()) {
   what <- spec(what, env = env)
 
   if (is_null(what$arg)) {
-    msg <- glue::glue_data(what, "{fn}() is {stage}")
+    msg <- sprintf("%s() is %s", what$fn, stage)
   } else {
-    msg <- glue::glue_data(what, "{fn}(arg) is {stage}")
+    msg <- sprintf("%s(%s) is %s", what$fn, what$arg, stage)
   }
 
   if (!is_null(with)) {