--- a/R/verbosity.R
+++ b/R/verbosity.R
@@ -52,13 +52,15 @@ lifecycle_verbosity <- function() {
 
   if (!is_string(opt, c("quiet", "default", "warning", "error"))) {
     options(lifecycle_verbosity = "default")
-    warn(glue::glue(
-      "
-      The `lifecycle_verbosity` option must be set to one of:
-      \"quiet\", \"default\", \"warning\", or \"error\".
-      Resetting to \"default\".
-      "
-    ))
+
+    message <- paste(
+      sep = " ",
+      "The `lifecycle_verbosity` option must be set to one of:",
+      "\"quiet\", \"default\", \"warning\", or \"error\".",
+      "Resetting to \"default\"."
+    )
+    
+    warn(message)
   }
 
   opt