The 3 commits here are self contained. Look at them individually to easily understand this PR.

The goal of this PR is to improve performance of `deprecate_soft()` and friends. We've been plagued by poor performance here for years, and it really hurts when used repeatedly in a `group_by()` + `mutate()` where the function used in `mutate()` is deprecated or has a deprecated argument.

I've been able to make it ~3x faster by:
- Swapping `glue_data()` for `sprintf()`, at the cost of ergonomics. But this is the biggest win, so I think it is worth it.
- Swapping `arg_match()` for `arg_match0()`
- Avoiding `-.POSIXct` and difftime operations when computing the 8 hour time diff, which was doing the wrong thing anyways!!

``` r
cross::bench_branches(\() {
  library(lifecycle)
  
  # trigger the 8 hour warning once
  deprecate_soft("1.1.0", I("my thing"), details = "for this reason")

  bench::mark(
    deprecate_soft("1.1.0", I("my thing"), details = "for this reason")
  )
})
#> # A tibble: 2 × 7
#>   branch               expression                    min  median `itr/sec` mem_alloc `gc/sec`
#>   <chr>                <bch:expr>                <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl>
#> 1 fix/deprecation-time "deprecate_soft(\"1.1.0\…  73.1µs  75.3µs    12802.    8.56KB     168.
#> 2 main                 "deprecate_soft(\"1.1.0\… 224.8µs 231.5µs     4229.    8.56KB     176.
```

I got here by analyzing this profvis output

```r
profvis::profvis(for (i in 1:100000) lifecycle::deprecate_soft("1.5.0", I("a thing"), details = "another thing"))
```

<img width="496" height="747" alt="Screenshot 2025-10-03 at 10 32 46 AM" src="https://github.com/user-attachments/assets/999929a0-0b08-443a-b0b9-bb97680485cb" />